requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
crispy-tailwind==1.0.3
Django==4.2.29
django-crispy-forms==2.5
djangorestframework==3.16.1
packaging==26.0
sqlparse==0.5.5
typing_extensions==4.15.0
